Remove aws_worlds (git submodule).
Given that the aws_world package is designed for ROS1, including it in the 'src' directory requires adding `COLCON_IGNORE` to prevent errors during colcon builds. However, this approach becomes inconvenient if aws_world isn't packaged, as it prevents the use of commands like `FindPackageShare` to access its path. Consequently, I've decided to address this by directly copying the model and world files in the upcoming commit.