Copy model from fuel_model for aws_hospital.
As stated in the README.md of aws_hospital, the model needs to be downloaded from fuel_model using the `setup.sh` script. However, upon inspection, I noticed that the URL for the fuel_model has been removed. The new URL is now https://fuel.gazebosim.org/1.0/OpenRobotics/models/.  Before sourcing the setup script, it's necessary to update this URL in `fuel_utility.py`. To avoid potential issues with downloading the model in the future, I've decided to commit these models.

Reference:
- https://github.com/aws-robotics/aws-robomaker-hospital-world/blob/7161eb8448f5cba6a469da7a79ae5d660b0b7f58/README.md
- https://app.gazebosim.org/fuel/models